+++ fractal_server/tasks/v2/local/reset.py
@@ -6,7 +6,7 @@ from fractal_server.app.db import get_sync_db
 from fractal_server.app.models import Profile
 from fractal_server.app.models import Resource
 from fractal_server.app.schemas.v2 import TaskGroupActivityAction
-from fractal_server.app.schemas.v2.task_group import TaskGroupActivityStatus
+from fractal_server.app.schemas.v2 import TaskGroupActivityStatus
 from fractal_server.logger import reset_logger_handlers
 from fractal_server.logger import set_logger
 from fractal_server.tasks.utils import get_log_path
@@ -25,7 +25,7 @@ from ._utils import _customize_and_run_template
 from ._utils import rmtree_nofail
 
 
-def reactivate_local(
+def reset_local(
     *,
     task_group_activity_id: int,
     task_group_id: int,
@@ -33,15 +33,21 @@ def reactivate_local(
     profile: Profile,
 ) -> None:
     """
-    Reactivate a task group venv.
+    Re-collect a task package.
 
-    This function is run as a background task, therefore exceptions must be
+    This function runs as a background task, therefore exceptions must be
     handled.
 
+    NOTE:  since this function is sync, it runs within a thread - due to
+    starlette/fastapi handling of background tasks (see
+    https://github.com/encode/starlette/blob/master/starlette/background.py).
+
+
     Args:
-        task_group_id:
         task_group_activity_id:
+        task_group_id:
         resource:
+        profile:
     """
 
     LOGGER_NAME = f"{__name__}.ID{task_group_activity_id}"
@@ -53,7 +59,7 @@ def reactivate_local(
             log_file_path=log_file_path,
         )
 
-        logger.debug("START")
+        logger.info("START")
         with next(get_sync_db()) as db:
             db_objects_ok, task_group, activity = get_activity_and_task_group(
                 task_group_activity_id=task_group_activity_id,
@@ -64,24 +70,7 @@ def reactivate_local(
             if not db_objects_ok:
                 return
 
-            # Check that the (local) task_group venv_path does not exist
-            if Path(task_group.venv_path).exists():
-                error_msg = f"{task_group.venv_path} already exists."
-                logger.error(error_msg)
-                fail_and_cleanup(
-                    task_group=task_group,
-                    task_group_activity=activity,
-                    logger_name=LOGGER_NAME,
-                    log_file_path=log_file_path,
-                    exception=FileExistsError(error_msg),
-                    db=db,
-                )
-                return
-
             try:
-                activity.status = TaskGroupActivityStatus.ONGOING
-                activity = add_commit_refresh(obj=activity, db=db)
-
                 # Prepare replacements for templates
                 python_bin = get_python_interpreter(
                     python_version=task_group.python_version,
@@ -92,61 +81,76 @@ def reactivate_local(
                     python_bin=python_bin,
                     resource=resource,
                 )
-                with open(f"{tmpdir}/pip_freeze.txt", "w") as f:
-                    f.write(task_group.env_info)
-                replacements.append(
-                    ("__PIP_FREEZE_FILE__", f"{tmpdir}/pip_freeze.txt")
-                )
-                # Prepare common arguments for `_customize_and_run_template`
+
+                # Prepare common arguments for `_customize_and_run_template``
                 common_args = dict(
                     replacements=replacements,
                     script_dir=(
                         Path(task_group.path) / SCRIPTS_SUBFOLDER
                     ).as_posix(),
                     prefix=(
-                        f"{int(time.time())}_"
-                        f"{TaskGroupActivityAction.REACTIVATE}"
+                        f"{int(time.time())}_{TaskGroupActivityAction.RESET}"
                     ),
                     logger_name=LOGGER_NAME,
                 )
 
-                logger.debug("start - create venv")
+                # Set status to ONGOING and refresh logs
+                activity.status = TaskGroupActivityStatus.ONGOING
+                activity.log = get_current_log(log_file_path)
+                activity = add_commit_refresh(obj=activity, db=db)
+
+                # Run script 1
                 _customize_and_run_template(
                     template_filename="1_create_venv.sh",
                     **common_args,
                 )
-                logger.debug("end - create venv")
                 activity.log = get_current_log(log_file_path)
                 activity = add_commit_refresh(obj=activity, db=db)
 
-                logger.debug("start - install from pip freeze")
+                # Run script 2
                 _customize_and_run_template(
-                    template_filename="5_pip_install_from_freeze.sh",
+                    template_filename="2_pip_install.sh",
                     **common_args,
                 )
-                logger.debug("end - install from pip freeze")
                 activity.log = get_current_log(log_file_path)
+                activity = add_commit_refresh(obj=activity, db=db)
+
+                # Run script 3
+                pip_freeze_stdout = _customize_and_run_template(
+                    template_filename="3_pip_freeze.sh",
+                    **common_args,
+                )
+                activity.log = get_current_log(log_file_path)
+                activity = add_commit_refresh(obj=activity, db=db)
+
+                # Update task_group data
+                logger.info("Update TaskGroupV2 - start")
+                task_group.env_info = pip_freeze_stdout
+                task_group.active = True
+                task_group = add_commit_refresh(obj=task_group, db=db)
+                logger.info("Update TaskGroupV2 - end")
+
+                # Finalize (write metadata to DB)
+                logger.info("finalising - START")
                 activity.status = TaskGroupActivityStatus.OK
                 activity.timestamp_ended = get_timestamp()
                 activity = add_commit_refresh(obj=activity, db=db)
-                task_group.active = True
-                task_group = add_commit_refresh(obj=task_group, db=db)
-                logger.debug("END")
+                logger.info("finalising - END")
+                logger.info("END")
 
                 reset_logger_handlers(logger)
 
-            except Exception as reactivate_e:
+            except Exception as collection_e:
                 rmtree_nofail(
                     folder_path=task_group.venv_path,
                     logger_name=LOGGER_NAME,
                 )
-
                 fail_and_cleanup(
                     task_group=task_group,
                     task_group_activity=activity,
                     logger_name=LOGGER_NAME,
                     log_file_path=log_file_path,
-                    exception=reactivate_e,
+                    exception=collection_e,
                     db=db,
                 )
         return