+++ fractal_server/tasks/v2/utils_background.py
@@ -76,6 +76,9 @@ def fail_and_cleanup(
         case TaskGroupActivityAction.DEACTIVATE:
             task_group.active = True
             db.add(task_group)
+        case TaskGroupActivityAction.RESET:
+            task_group.active = False
+            db.add(task_group)
     db.commit()
     reset_logger_handlers(logger)